CiCaWa-HashITup/backend/requirements.txt

fastapi>=0.117.1
google-genai>=1.39.1
passlib>=1.7.4
pydantic>=2.11.9
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20
python-socketio>=5.13.0
sqlalchemy>=2.0.43
uvicorn>=0.37.0
websockets>=15.0.1
